PluginFeatureDetection and disable WBOIT in Safari
